remove search code & fix routes
