+++ ui/src/components/General/Search/search.module.scss
@@ -1,0 +1,12 @@
+.search {
+	width: var(--bt-list-card-container);
+	margin: 0 auto;
+	margin-bottom: var(--bt-size-10);
+
+	input {
+		margin-bottom: var(--bt-size-10);
+	}
+	&__category {
+		cursor: pointer;
+	}
+}
