+++ ui/src/components/Hoc/List/List.tsx
@@ -1,13 +1,9 @@
-import { ChangeEvent, useState, useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
-import { Input } from "antd";
 import style from "./list.module.scss";
 import { ListProps } from "./types";
 import { NewsType } from "components/General/ListItems/News/news.types";
 import { ResourceType } from "components/General/ListItems/Resource/resource.type";
-import { getCategories } from "./helper";
 import { QUERY_KEY_NEWS } from "pages/News";
-import CategoryTags from "./CategoryTags/CategoryTags";
 
 const filteredNews = (searchQuery: string, items: NewsType[]) => {
 	if (searchQuery) {
@@ -50,29 +46,11 @@ const List = <T,>({
 	isLoading,
 	isError,
 }: ListProps<T>): JSX.Element => {
-	const [searchParams, setSearchParams] = useSearchParams();
-	const [queryParams, setQueryParams] = useState({
+	const [searchParams] = useSearchParams();
+
+	const { q: searchQuery, cat: categoryQuery } = {
 		q: searchParams.get("q") || "",
 		cat: searchParams.get("cat") || "",
-	});
-
-	const { q: searchQuery, cat: categoryQuery } = queryParams;
-
-	useEffect(() => {
-		if (resourceName === QUERY_KEY_NEWS) {
-			setSearchParams(`?q=${searchQuery}`);
-		} else {
-			setSearchParams(`?q=${searchQuery}&cat=${categoryQuery}`);
-		}
-	}, [searchQuery, categoryQuery, setSearchParams, resourceName]);
-
-	const handleSearchChange = (e: ChangeEvent<HTMLInputElement>) => {
-		const value = e.target.value;
-		setQueryParams((prevParams) => ({ ...prevParams, q: value }));
-	};
-
-	const handleCategoryChange = (value: string) => {
-		setQueryParams((prevParams) => ({ ...prevParams, cat: value }));
 	};
 
 	const filteredList =
@@ -88,8 +66,6 @@ const List = <T,>({
 		window.open(url, "_blank");
 	};
 
-	const categories = getCategories(items as ResourceType[], resourceName);
-
 	const list = filteredList ? filteredList : [...Array(20).keys()];
 	if (isError) {
 		return <div>Something went wrong</div>;
@@ -97,20 +73,6 @@ const List = <T,>({
 
 	return (
 		<div className={style.container}>
-			<Input
-				type="text"
-				placeholder="Search by title..."
-				value={searchQuery}
-				onChange={handleSearchChange}
-			/>
-
-			<CategoryTags
-				categories={categories}
-				category={categoryQuery}
-				handleCategoryChange={handleCategoryChange}
-				className={style.container__tag}
-			/>
-
 			{list.map((item, i) => (
 				<ItemComponent
 					key={i}
