+++ ui/src/pages/List/Book/Book.tsx
@@ -1,6 +1,7 @@
 import List from "components/Hoc/List/List";
 import Resource from "components/General/ListItems/Resource/Resource";
 import useFetchList from "lib/utils/hooks/useFetchList";
+import Search from "components/General/Search/Search";
 
 const URL = `./book.json`;
 const QUERY_KEY = "book";
@@ -9,13 +10,16 @@ const Book: React.FC = () => {
 	const { data, isLoading, isError } = useFetchList(QUERY_KEY, URL);
 
 	return (
-		<List
-			items={data}
-			resourceName={QUERY_KEY}
-			itemComponent={Resource}
-			isLoading={isLoading}
-			isError={isError}
-		/>
+		<>
+			<Search items={data} resourceName={QUERY_KEY} />
+			<List
+				items={data}
+				resourceName={QUERY_KEY}
+				itemComponent={Resource}
+				isLoading={isLoading}
+				isError={isError}
+			/>
+		</>
 	);
 };
